Use custom.el to customize env
Now user no longer needs to manually edit env.el.
When emacs starts for the first time, user will be
prompted to configure that. Those configuration will
be saved as customized variables in custom.el.

So env-default.el and env.el are no longer used.
Now env-default is kept for a while, until ede project
config can be automated in the future